% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex2img.R
\name{tex2img}
\alias{tex2img}
\title{Convert TeX equations to images using Sciweavers iTex2Img}
\usage{
tex2img(
  tex,
  color = "black",
  background = "transparent",
  font = "arev",
  size = 8,
  type = "png"
)
}
\arguments{
\item{tex}{A character string of TeX code. Be sure to escape backslashes.}

\item{color}{A character string for the text color.
Use \link{tex2img_colors} to see options.}

\item{background}{A character string for the background color.
Use \link{tex2img_backgrounds} to see options.}

\item{font}{A character string for the font face.
Use \link{tex2img_fonts} to see options.}

\item{size}{A positive number for the font size.}

\item{type}{A character string for the output file type.
Use \link{tex2img_types} to see options.}
}
\value{
The Markdown code to embed an image from Sciweavers iTex2Img
}
\description{
Convert TeX equations to images using Sciweavers iTex2Img
}
\examples{
tex2img("\\\\frac{22}{7}")

tex2img(
  "\\\\frac{355}{113}",
  color = "Blue", background = "Yellow",
  font = "Modern", size = 18, type = "jpg"
)
}
