% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tex2img_colors}
\alias{tex2img_colors}
\alias{tex2img_backgrounds}
\alias{tex2img_fonts}
\alias{tex2img_types}
\title{tex2img options}
\format{
Character vectors:
\describe{\item{\code{tex2img_colors}}{valid text colors for \code{color}}}

\describe{
\item{\code{tex2img_backgrounds}}{valid background colors for \code{background}}
}

\describe{\item{\code{tex2img_fonts}}{valid font faces for \code{font}}}

\describe{\item{\code{tex2img_types}}{valid output file types for \code{type}}}
}
\source{
\url{http://www.sciweavers.org/free-online-latex-equation-editor}
}
\usage{
tex2img_colors

tex2img_backgrounds

tex2img_fonts

tex2img_types
}
\description{
Character vectors listing the valid options for the arguments in \code{\link[=tex2img]{tex2img()}}
}
\keyword{datasets}
